+++ x/leverage/keeper/borrows.go
@@ -57,8 +57,8 @@ func (k Keeper) GetAvailableToBorrow(ctx sdk.Context, denom string) sdk.Int {
 	return sdk.MaxInt(moduleBalance.Sub(reserveAmount), sdk.ZeroInt())
 }
 
-// DeriveBorrowUtilization calculates the current supply utilization of a token denom.
-func (k Keeper) DeriveBorrowUtilization(ctx sdk.Context, denom string) sdk.Dec {
+// SupplyUtilization calculates the current supply utilization of a token denom.
+func (k Keeper) SupplyUtilization(ctx sdk.Context, denom string) sdk.Dec {
 	// Supply utilization is equal to total borrows divided by the token supply
 	// (including borrowed tokens yet to be repaid and excluding tokens reserved).
 	moduleBalance := k.ModuleBalance(ctx, denom).ToDec()